Huge update - unfortunately, was unable to commit individually
Added the rotation to pacman so that pacman is facing in the direction that its moving in. I also rotated in along various axis so that the pacman mouth can be noticed

Fix my frame callback so that the perspective of the objects don't change.

Added the code for the mouse movements. Not really working correctly though